Object Detection and Measure Distance using YOLOv8/requirements.txt
opencv-python
ultralytics
supervision
pytorch
torchvision